Split Path extension into a individual file
